pyproject.toml
[project]
name = "pi-rss-publisher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.3",
    "feedgen>=1.0.0",
    "feedparser>=6.0.12",
    "psycopg[binary,pool]>=3.2.12",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
